protect routes middleware on backend implemented

+++ frontend/src/components/WorkoutDetails.jsx
@@ -1,14 +1,23 @@
-import { FaTimes } from 'react-icons/fa'
+// Context
 import { useWorkoutsContext } from '../hooks/useWorkoutsContext'
+import { useAuthContext } from '../hooks/useAuthContext';
 // Date FNS
 import formatDistanceToNow from 'date-fns/formatDistanceToNow';
+// React Icons
+import { FaTimes } from 'react-icons/fa'
 
 export const WorkoutDetails = ({ workout }) => {
     const { dispatch } = useWorkoutsContext();
+    const { user } = useAuthContext();
 
     const deleteWorkout = async () => {
+      if(!user) return
+
       const res = await fetch('/api/workouts/' + workout._id, {
-        method: 'DELETE'
+        method: 'DELETE',
+        headers: {
+          'Authorization': `Bearer ${user.token}`
+        }
       })
       const data = await res.json();
